requirements.txt
streamlit
langchain
faiss-cpu
pypdf
python-dotenv
langchain_google_genai